+++ src/components/Home-Page/JobsList/jobsListStyle.ts
