+++ src/components/Personal-Zone/Employer/JobsList/jobsListStyle.ts
