+++ src/components/Personal-Zone/Employer/header/PostNewJob/JobNicknameInput.tsx
