+++ src/components/Personal-Zone/Employer/header/PostNewJob/JobSeniorityInput.tsx
