+++ src/components/Personal-Zone/Employer/header/PostNewJob/PostNewJobStyles.ts
