+++ src/components/Personal-Zone/Employer/header/PostNewJob/SwitchInput.tsx
